+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 0.2.0
 -----
 * Switch to HTTPS jQuery
+* Added h-card support
 
 0.1.0
 -----